Fix problem when using 0 seconds for stop timeout
It seems as if this was caused by an update to one of the dependencies (maybe anyio?!). The CI pipeline did still pass for Linux, but raised some warning about the event loop already being closed on the worker thread.

+++ python_packages/jupyter_lsp/jupyter_lsp/session.py
@@ -206,9 +206,12 @@ class LanguageServerSessionBase(
         if self.process is None:  # pragma: no cover
             return
 
+        if timeout < 0.0:  # pragma: no cover
+            raise ValueError("Timeout must not be negative!")
+
         # try to stop the process gracefully
         self.process.terminate()
-        with anyio.move_on_after(timeout):
+        with anyio.move_on_after(timeout + 0.1):  # avoid timeout of 0s
             self.log.debug("Waiting for process to terminate")
             await self.process.wait()
             return
